+++ data/duck.txt
@@ -1,0 +1,3 @@
+E | 0 | project meeting | Sep 08 2032 12:30-Sep 08 2032 12:30
+D | 0 | return book | Apr 20 2023 14:40
+D | 0 | return book | Aug 09 2022 14:40
