+++ src/main/java/duck/Storage.java
@@ -3,7 +3,6 @@ package duck;
 import duck.task.*;
 import duck.exceptions.FileIOException;
 
-
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -40,27 +39,27 @@ public class Storage {
             }
             FileWriter fw = new FileWriter(FILE);
             for (int i = 0; i < list.size(); i++) {
-               Task task = list.get(i);
-               switch (task.type()) {
-                   case "D":
-                       Deadline deadline = (Deadline) task;
-                       fw.write(deadline.type() + " | " + (deadline.getStatusIcon().isBlank() ? "0" : "1") + " | " +
-                               deadline.getDescription() + " | " +
-                               deadline.getBy());
-                       break;
-                   case "E":
-                       Events events = (Events) task;
-                       fw.write(events.type() + " | " + (events.getStatusIcon().isBlank() ? "0" : "1") + " | " +
-                               events.getDescription() + " | " +
-                               events.getDate());
-                       break;
-                   case "T":
-                       ToDo toDo = (ToDo) task;
-                       fw.write(toDo.type() + " | " + (toDo.getStatusIcon().isBlank() ? "0" : "1") + " | " +
-                               toDo.getDescription());
-                       break;
-               }
-               fw.write("\n");
+                Task task = list.get(i);
+                switch (task.type()) {
+                    case "D":
+                        Deadline deadline = (Deadline) task;
+                        fw.write(deadline.type() + " | " + (deadline.getStatusIcon().isBlank() ? "0" : "1") + " | " +
+                                deadline.getDescription() + " | " +
+                                deadline.getBy());
+                        break;
+                    case "E":
+                        Events events = (Events) task;
+                        fw.write(events.type() + " | " + (events.getStatusIcon().isBlank() ? "0" : "1") + " | " +
+                                events.getDescription() + " | " +
+                                events.getDate());
+                        break;
+                    case "T":
+                        ToDo toDo = (ToDo) task;
+                        fw.write(toDo.type() + " | " + (toDo.getStatusIcon().isBlank() ? "0" : "1") + " | " +
+                                toDo.getDescription());
+                        break;
+                }
+                fw.write("\n");
             }
             fw.close();
         } catch (IOException e) {
@@ -76,12 +75,13 @@ public class Storage {
             }
             FileReader fr = new FileReader(FILE);
             Scanner sc = new Scanner(fr);
-            while(sc.hasNext()) {
+            while (sc.hasNext()) {
                 String line = sc.nextLine();
                 String[] split = line.split(" \\| ");
                 switch (split[0]) {
                     case "D":
-                        Deadline tempDeadline = new Deadline(split[2], LocalDateTime.parse(split[3], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")));
+                        Deadline tempDeadline = new Deadline(split[2],
+                                LocalDateTime.parse(split[3], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")));
                         if (split[1].equals("1")) {
                             tempDeadline.markAsDone();
                         }
@@ -89,7 +89,10 @@ public class Storage {
                         break;
                     case "E":
                         String[] start_end = split[3].split("-");
-                        Events tempEvent = new Events(LocalDateTime.parse(start_end[0], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")), LocalDateTime.parse(start_end[1], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")), split[2]);
+                        Events tempEvent = new Events(
+                                LocalDateTime.parse(start_end[0], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")),
+                                LocalDateTime.parse(start_end[1], DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")),
+                                split[2]);
                         if (split[1].equals("1")) {
                             tempEvent.markAsDone();
                         }
