+++ text-ui-test/data/duck.txt
@@ -1,0 +1,2 @@
+D | 0 | return book | Mar 09 2034 12:00
+E | 0 | project meeting | Sept 08 2032 12:30-Sept 08 2032 12:30
